+++ package.json
@@ -55,8 +55,8 @@
     "lint": "eslint src && echo '-- linting finished --'",
     "build": "babel src -d dist",
     "prettier": "prettier --write 'src/**/*.js'",
-    "test": "tape -r babel-register -r babel-polyfill src/test",
-    "test:integration": "tape -r babel-register -r babel-polyfill src/integration",
+    "test": "tape -r babel-register -r babel-polyfill test/unit.js",
+    "test:integration": "tape -r babel-register -r babel-polyfill test/integration",
     "watch": "watch 'reset && npm run --silent test' src",
     "precommit-msg": "echo 'Pre-commit checks...' && exit 0",
     "pretty-quick": "pretty-quick --staged",
