--- test/index.js
@@ -1,4 +1,0 @@
-import './configure-feature.test';
-import '../feature-toggles.test';
-import '../feature.test';
-import '../with-feature-toggles.test';
