+++ test/unit.js
@@ -1,0 +1,4 @@
+import '../src/configure-feature.test';
+import '../src/feature-toggles.test';
+import '../src/feature.test';
+import '../src/with-feature-toggles.test';
